update:bookmarkページにコンポーネント分割
