+++ front/src/pages/job/bookmark.tsx
@@ -1,63 +1,12 @@
 import Head from "next/head";
 import Layout from "@/features/jobs/Layout";
 import Loading from "@/components/Loading";
-import {
-  jobBookmarkAtom,
-  jobBookmarkIdsAtom,
-  loadingAtom,
-} from "@/atoms/atoms";
-import { Box, Typography } from "@mui/material";
-import { useEffect } from "react";
-import { useAtom, useSetAtom } from "jotai";
-import JobList from "@/features/jobs/JobList";
-import fetch from "@/libs/fetch";
-import { JobListResponse } from "@/features/jobs/job.type";
-import BasicPagination from "@/components/BasicPagination";
-import { useRouter } from "next/router";
-import LoadPage from "@/components/LoadPage";
+import { loadingAtom } from "@/atoms/atoms";
+import { useAtomValue } from "jotai";
+import BookmarkContent from "@/features/jobs/bookmark/BookmarkContent";
 
 export default function Bookmark() {
-  const [isLoading, setLoading] = useAtom(loadingAtom);
-  const setBookMarkIds = useSetAtom(jobBookmarkIdsAtom);
-  const [bookmark, setBookMark] = useAtom(jobBookmarkAtom);
-  const router = useRouter();
-
-  const handleChangePagination = async (
-    _: React.ChangeEvent<unknown>,
-    value: number
-  ) => {
-    try {
-      setLoading(true);
-      const { data } = await fetch.get<JobListResponse>(
-        `/api/v1/jobBookmarkList?page=${value}`
-      );
-      setBookMark(data);
-      router.push(`/job/bookmark?page=${value}`, undefined, {
-        shallow: true,
-      });
-    } catch (error) {
-      console.error(error);
-    } finally {
-      setLoading(false);
-      window.scrollTo({ top: 0 });
-    }
-  };
-
-  useEffect(() => {
-    (async () => {
-      setLoading(true);
-      const { data: bookmark } = await fetch.get<JobListResponse>(
-        "/api/v1/jobBookmarkList"
-      );
-      const { data: bookmarkIds } = await fetch.get<number[]>(
-        `/api/v1/jobBookmarkIds`
-      );
-      setBookMark(bookmark);
-      setBookMarkIds(Object.values(bookmarkIds));
-    })().finally(() => {
-      setLoading(false);
-    });
-  }, [setLoading, setBookMarkIds, setBookMark]);
+  const isLoading = useAtomValue(loadingAtom);
 
   return (
     <>
@@ -69,36 +18,7 @@ export default function Bookmark() {
       </Head>
       <div>
         <Layout>
-          <Typography fontSize={24} pb={2} fontWeight={500} variant="h2">
-            ブックマークした求人
-          </Typography>
-          {isLoading ? (
-            <LoadPage />
-          ) : bookmark && bookmark.jobList.length ? (
-            <>
-              <JobList
-                jobList={bookmark.jobList}
-                showBookmark={false}
-                showDeleteBookmark
-              />
-              <Box pt={2} pb={8}>
-                <BasicPagination
-                  jobData={bookmark}
-                  handleChange={handleChangePagination}
-                />
-              </Box>
-            </>
-          ) : (
-            <Typography
-              textAlign={"center"}
-              fontSize={20}
-              fontWeight={600}
-              pt={4}
-              pb={20}
-            >
-              ブックマークがありません。
-            </Typography>
-          )}
+          <BookmarkContent />
         </Layout>
       </div>
       <Loading open={isLoading} />
